refactor read and write tests under spark module
