+++ java/spark/src/test/java/com/lancedb/lance/spark/read/SparkConnectorReadWithRowAddress.java
@@ -13,70 +13,14 @@
  */
 package com.lancedb.lance.spark.read;
 
-import com.lancedb.lance.spark.LanceConfig;
-import com.lancedb.lance.spark.LanceDataSource;
 import com.lancedb.lance.spark.TestUtils;
 
-import org.apache.spark.sql.Dataset;
-import org.apache.spark.sql.Row;
-import org.apache.spark.sql.SparkSession;
-import org.junit.jupiter.api.AfterAll;
-import org.junit.jupiter.api.BeforeAll;
 import org.junit.jupiter.api.Test;
 
 import java.util.Arrays;
-import java.util.List;
 import java.util.stream.Collectors;
 
-import static org.junit.jupiter.api.Assertions.assertEquals;
-
-public class SparkConnectorReadWithRowAddress {
-  private static SparkSession spark;
-  private static String dbPath;
-  private static Dataset<Row> data;
-
-  @BeforeAll
-  static void setup() {
-    spark =
-        SparkSession.builder()
-            .appName("spark-lance-connector-test")
-            .master("local")
-            .config("spark.sql.catalog.lance", "com.lancedb.lance.spark.LanceCatalog")
-            .getOrCreate();
-    dbPath = TestUtils.TestTable1Config.dbPath;
-    data =
-        spark
-            .read()
-            .format(LanceDataSource.name)
-            .option(
-                LanceConfig.CONFIG_DATASET_URI,
-                LanceConfig.getDatasetUri(dbPath, TestUtils.TestTable1Config.datasetName))
-            .load();
-  }
-
-  @AfterAll
-  static void tearDown() {
-    if (spark != null) {
-      spark.stop();
-    }
-  }
-
-  private void validateData(Dataset<Row> data, List<List<Long>> expectedValues) {
-    List<Row> rows = data.collectAsList();
-    assertEquals(expectedValues.size(), rows.size());
-
-    for (int i = 0; i < rows.size(); i++) {
-      Row row = rows.get(i);
-      List<Long> expectedRow = expectedValues.get(i);
-      assertEquals(expectedRow.size(), row.size());
-
-      for (int j = 0; j < expectedRow.size(); j++) {
-        long expectedValue = expectedRow.get(j);
-        long actualValue = row.getLong(j);
-        assertEquals(expectedValue, actualValue, "Mismatch at row " + i + " column " + j);
-      }
-    }
-  }
+public class SparkConnectorReadWithRowAddress extends SparkConnectorReadTestBase {
 
   @Test
   public void readAllWithoutRowAddr() {
